Fix repeating weekday runs
For some godforsaken reason, Google Calendar will only repeat an event
up to 700ish times. Therefore, we can't just start the repeating weekday
run event in the distant past and have everything workout.

Instead, this starts the weekday runs on the most previous Tuesday
before today (when the iCal is generated) or the first run in the
schedule, whichever is earlier.