Implement basic JSON-as-flatfile database for photo urls
Google Photos URLs have gotten much longer, and we have
many more of them. The index page is a lot of HTML now.
This is bad for browsers and for search engines.
Dynamically fetching random chunks of the photo url file
is a workaround until we have a better solution where
the URLs can follow a well-known format.
Falls back to fetching the whole JSON if the browser can't
make range requests, but will still only add images to the
DOM incrementally