build routes.yml from route-db.csv
For now this works by moving the old routes.yml to legacy-routes.yml
and concatenating that onto the YAML produced after normalizing
route-db.csv and converting to YAML.

Makefiles updated to orchaestrate these pieces.

I do not love that the route-db-invariants.py script knows where
everything lives, BUT it is essentially a build script so I think
it is OK in this case. Taking the paths as an argument could work,
but it also needs to know where to check for GPX files etc.

Finally, also had to update the Gemfile.lock to get the site to
build. Not sure what the consequences of that may be.