use CSS variable to fix linting error
This unlocks the VS Code autoformatter which is a bit of a double-edged
sword. For now I have left off the autoformatter b/c it is a big change.
I also don't like it b/c auto formatters / pretty printers are not smart
enough to appreciate and maintain delicate alignment across lines that
preservers / improves readability.